+++ OrangeDesignSystem/Sources/OrangeDesignSystem/Modules/AboutView.swift
@@ -41,6 +41,7 @@ public struct AboutView: View {
                 applicationDescription.imageHeader
                     .resizable()
                     .aspectRatio(contentMode: .fit)
+                    .accessibilityHidden(true)
 
                 ApplicationDescriptionView()
                     .padding(.all, ODSSpacing.m)
@@ -179,20 +180,30 @@ private struct ApplicationDescriptionView: View {
 
     var body: some View {
         VStack(alignment: .leading, spacing: 4) {
-            Text(applicationDescription.applicationName).odsFont(.largeTitle)
+            Text(applicationDescription.applicationName)
+                .odsFont(.largeTitle)
+                .fixedSize(horizontal: false, vertical: true)
+
             Text("Version \(applicationDescription.applicationVersion)")
+                .fixedSize(horizontal: false, vertical: true)
+
             if let buildNumber = applicationDescription.applicationBuildNumber {
                 Text("Build \(buildNumber)")
+                    .fixedSize(horizontal: false, vertical: true)
             }
+
             if let buildType = applicationDescription.applicationBuildType {
                 Text(buildType)
+                    .fixedSize(horizontal: false, vertical: true)
             }
 
             Text(applicationDescription.copyrightNotice)
+                .fixedSize(horizontal: false, vertical: true)
         }
     }
 }
 
+#if DEBUG
 struct AboutView_Previews: PreviewProvider {
     static var previews: some View {
         let applicationDescription = ApplicationDescription(applicationName: "APP NAME", applicationVersion: "1.0.0", applicationBuildNumber: "123456789", applicationBuildType: "PREVIEW")
@@ -205,3 +216,4 @@ struct AboutView_Previews: PreviewProvider {
         }
     }
 }
+#endif