+++ OrangeDesignSystemDemo/OrangeDesignSystemDemo/Views/Components/Pages/ProgressIndicatorPage.swift
@@ -24,33 +24,31 @@
 import OrangeDesignSystem
 import SwiftUI
 
-struct ProgressIndicatorPage: View {
+struct ProgressIndicatorComponent: Component {
+    let title: String
+    let image: Image
+    let description: String
+    let variants: AnyView
+    
+    init() {
+        title = "Progress indicators"
+        image = Image("Progress_indicator")
+        description =  "Progress indicators show users that elements or pages are loading."
+        
+        variants = AnyView(ProgressIndicatorVariants())
+    }
+}
 
+struct ProgressIndicatorVariants: View {
     var body: some View {
-        ScrollView {
-            Image("Progress_indicator")
-                .resizable()
-                .aspectRatio(contentMode: .fill)
-
-            VStack(alignment: .leading, spacing: ODSSpacing.l) {
-                ComponentDescription(text: "Progress indicators show users that elements or pages are loading.")
-                VariantsTitle()
-                ProgressBar()
-                ProgressIndicator()
-            }
-            .padding(EdgeInsets(top: ODSSpacing.none, leading: ODSSpacing.m, bottom: ODSSpacing.xs, trailing: ODSSpacing.m))
+        VStack(alignment: .leading, spacing: ODSSpacing.m) {
+            ProgressBarVariant()
+            ProgressIndicatorVariant()
         }
-        .background(ODSColor.primaryBackground.color)
     }
 }
 
-struct ProgressIndicator_Previews: PreviewProvider {
-    static var previews: some View {
-        ProgressIndicator()
-    }
-}
-
-struct ProgressBar: View {
+struct ProgressBarVariant: View {
     @State private var secondsElapsed = 0.0
     @State private var timer = Timer.publish(every: 0.1, on: .main, in: .common).autoconnect()
     let maxSeconds: CGFloat = 100.0
@@ -69,7 +67,7 @@ struct ProgressBar: View {
     }
 }
 
-struct ProgressIndicator: View {
+struct ProgressIndicatorVariant: View {
     var body: some View {
         Text("Progress indicator")
             .odsFont(.title2)
@@ -80,3 +78,17 @@ struct ProgressIndicator: View {
         }
     }
 }
+
+#if DEBUG
+struct ProgressIndicatorPage_Previews: PreviewProvider {
+    static var previews: some View {
+        ForEach(ColorScheme.allCases, id: \.self) {
+            List {
+                ProgressIndicatorVariants()
+            }
+            .accentColor(ODSColor.coreOrange.color)
+            .preferredColorScheme($0)
+        }
+    }
+}
+#endif